tratamento de erros divisao de metodos para request/error
