+++ TestSantander-IBM/TestSantander-IBM/Scenes/Login/LoginModels.swift
@@ -37,7 +37,6 @@ struct LoginScene {
     }
     
     struct SelectUser {
-        
         struct Request {
         }
     }
